Remove encrypted-chunked feature
This patch always enables the syscalls that previously were behind the
encrypted-chunked feature.  This makes sure that enabling the feature in
one crate does not break another crate that also depends on
trussed-chunked.  In practice, the feature is always enabled anyway so
separating the encrypted syscalls does not bring any benefits.

Fixes: https://github.com/trussed-dev/trussed-staging/issues/20