Import HkdfExtension from Nitrokey/trussed-hkdf-backend
The HkdfExtension was previously maintained in a separate repository and
together with a custom backend.  Every additional backend adds some
overhead, both in the firmware and for maintenance.  Therefore this
patch moves the trussed-hkdf crate with the HkdfExtension as an
extension into this repository and implements it for the StagingBackend,
replacing the HkdfBackend.

This patch also releases trussed-hkdf v0.2.0 and trussed-staging v0.3.0.

Fixes: https://github.com/Nitrokey/trussed-hkdf-backend/issues/6